requirements.txt

# Microsoft Agent Framework (Preview)
agent-framework-azure-ai --pre

# Azure dependencies
azure-identity>=1.25.0

# Core dependencies
python-dotenv>=1.0.0
asyncio>=3.4.3

# Optional: For document generation
markdown>=3.5.0
weasyprint>=60.0

# Optional: For testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
